Refactor content schema helpers
Manual testing: not run (not available in container)
